Pokedex okay
