+++ pokedex/src/PokedexContext.js
@@ -1,0 +1,9 @@
+import React from "react";
+
+const PokedexContext = React.createContext({
+    pokemons: [],
+    removePokedex: (index) => { },
+    addPokedex: (index) => { }
+});
+
+export default PokedexContext;
